services/web/requirements.txt
boto3==1.24.12
flasgger==0.9.5
flask==2.0.3
flask-jwt-extended==4.4.1
flask-migrate==3.1.0
flask-sqlalchemy==2.5.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
python-json-logger==2.0.2
setuptools==62.1.0
sqlalchemy-utils==0.38.2